Front-end working on #108.
